custom fields: add divisor fields
At some point we need to split the difference between the ordered amount
and the amount we can order between the customers. For this to work, we
need to be able to know the minimum amount of items people need to buy
(sale amount), and the amount of stuff we can add or subtract from an
order. Obviously we need to start doing smart stuff when storing these
orders later on.